# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Movies",
    "version": "1.0.0",
    "title": "Casting Agency - FSND Capstone App",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/api/movies": {
      "get": {
        "tags": [
          "Movie Request"
        ],
        "summary": "Returns MovieList",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/MovieList"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Movie Request"
        ],
        "summary": "Add Movie",
        "requestBody": {
          "description": "Movie Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieRequestPostBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Movie"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
...
    "/api/actors": {
      "get": {
        "tags": [
          "Actor Request"
        ],
        "summary": "Returns ActorList",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/ActorList"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Actor Request"
        ],
        "summary": "Add Actor",
        "requestBody": {
          "description": "Actor Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActorRequestPostBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/Actor"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MovieRequestPostBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "format": "string"
          },
          "release_date": {
            "type": "number"
          }
        }
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "format": "string"
          },
          "releasedate": {
            "type": "number"
          },
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Actor"
            }
          }
        }
      },
      "MovieList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Movie"
            }
          }
        }
      },
      "Actor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          }
        }
      },
      "ActorList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Actor"
            }
          }
        }
      },
      "ActorRequestPostBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "number"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          }
        }
      },
      "Gender": {
        "type": "string",
        "enum": [
          "M", "F", "U"
        ]
      },
...
    }
  }
}